move proto/detect -> ssz/encoding/detect
