+++ encoding/ssz/detect/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
         "configfork.go",
         "fieldspec.go",
     ],
-    importpath = "github.com/prysmaticlabs/prysm/proto/detect",
+    importpath = "github.com/prysmaticlabs/prysm/encoding/ssz/detect",
     visibility = ["//visibility:public"],
     deps = [
         "//beacon-chain/state:go_default_library",
