+++ encoding/ssz/detect/configfork_test.go
@@ -90,7 +90,7 @@ func TestByState(t *testing.T) {
 		require.NoError(t, st.SetSlot(c.slot))
 		m, err := st.MarshalSSZ()
 		require.NoError(t, err)
-		cf, err := ByState(m)
+		cf, err := FromState(m)
 		require.NoError(t, err)
 		require.Equal(t, c.version, cf.Fork)
 		require.Equal(t, c.forkversion, cf.Version)
@@ -154,7 +154,7 @@ func TestUnmarshalState(t *testing.T) {
 		require.NoError(t, st.SetSlot(c.slot))
 		m, err := st.MarshalSSZ()
 		require.NoError(t, err)
-		cf, err := ByState(m)
+		cf, err := FromState(m)
 		require.NoError(t, err)
 		s, err := cf.UnmarshalBeaconState(m)
 		require.NoError(t, err)
@@ -260,7 +260,7 @@ func TestUnmarshalBlock(t *testing.T) {
 			b := c.b(t, c.slot)
 			marshaled, err := b.MarshalSSZ()
 			require.NoError(t, err)
-			cf, err := ByVersion(c.version)
+			cf, err := FromForkVersion(c.version)
 			require.NoError(t, err)
 			bcf, err := cf.UnmarshalBeaconBlock(marshaled)
 			if c.err != nil {
