+++ encoding/ssz/detect/fieldspec.go
