+++ encoding/ssz/detect/fieldspec_test.go
